pyproject.toml
[tool.poetry]
name = "llm-research"
version = "0.1.0"
description = "A minimum Python package built on top of the LangChain framework to interact with LLM."
authors = ["githubjacky"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
orjson = "^3.9.12"
tqdm = "^4.66.1"
langchain = "^0.1.3"
mlflow = "^2.9.2"
langchain-google-vertexai = "^0.0.2"
langchain-openai = "^0.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
pytest = "^7.4.4"
jupyterlab-execute-time = "2.3.1"
jupyterlab = "3.6.6"
jupyterlab-lsp = "4.3.0"
python-lsp-server = {extras = ["all"], version = "^1.8.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
